refactored firehose client out of options
