+++ src/Serilog.Sinks.Amazon.Kinesis.Firehose/Serilog.Sinks.Amazon.Kinesis.Firehose.csproj
@@ -67,6 +67,7 @@
     <Compile Include="KinesisFirehoseLoggerConfigurationExtensions.cs" />
     <Compile Include="Sinks\DurableKinesisFirehoseSink.cs" />
     <Compile Include="Sinks\HttpLogShipper.cs" />
+    <Compile Include="Sinks\HttpLogShipperBase.cs" />
     <Compile Include="Sinks\KinesisFirehoseSinkOptions.cs" />
     <Compile Include="Sinks\KinesisSinkState.cs" />
     <Compile Include="Sinks\KinesisFirehoseSink.cs" />
