+++ src/Serilog.Sinks.Amazon.Kinesis.Firehose/Sinks/DurableKinesisFirehoseSink.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using Amazon.KinesisFirehose;
 using Serilog.Core;
 using Serilog.Events;
 using Serilog.Sinks.RollingFile;
@@ -24,9 +25,9 @@ namespace Serilog.Sinks.Amazon.Kinesis.Firehose
         readonly HttpLogShipper _shipper;
         readonly RollingFileSink _sink;
 
-        public DurableKinesisFirehoseSink(KinesisFirehoseSinkOptions options)
+        public DurableKinesisFirehoseSink(KinesisFirehoseSinkOptions options, IAmazonKinesisFirehose kinesisFirehoseClient)
         {
-            var state = KinesisSinkState.Create(options);
+            var state = new KinesisSinkState(options, kinesisFirehoseClient);
 
             if (string.IsNullOrWhiteSpace(options.BufferBaseFilename))
             {
